requirements.txt
Flask>=2.0
requests>=2.28
beautifulsoup4>=4.12
pandas>=2.0
python-dateutil>=2.8

# optional for later
# mlflow
# sqlalchemy